pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools-scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "odsbox-jaquel-mcp"
dynamic = ["version"]
description = "ASAM ODS Jaquel MCP Server - A Model Context Protocol server for creating and validating Jaquel queries"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "Apache-2.0"}
authors = [
    {name = "Andreas K", email = "totonga@gmail.com"}
]
keywords = [
    "mcp",
    "jaquel",
    "odsbox",
    "asam",
    "ods",
    "measurement",
    "data access",
    "data science",
    "data analysis"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Scientific/Engineering :: Interface Engine/Protocol Translator",
    "Topic :: System :: Monitoring",
    "Topic :: Database :: Front-Ends",
]
dependencies = [
    "mcp>=0.1.0",
    "odsbox>=1.0.15",
    "pip-system-certs",
    "matplotlib",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=1.2.0",
    "black>=23.0",
    "isort>=5.0",
    "flake8>=6.0",
    "mypy>=1.0",
    "build>=0.8.0",
    "pandas-stubs>=2.3",
]
play = [
    "flask",
    "plotly",
    "seaborn",
    "scipy",
    "PyQt5",
    "jupyter",
]

[project.scripts]
odsbox-jaquel-mcp = "odsbox_jaquel_mcp.__main__:main"

[project.urls]
Homepage = "https://github.com/totonga/odsbox-jaquel-mcp/tree/main#readme"
Documentation = "https://github.com/totonga/odsbox-jaquel-mcp/tree/main#readme"
Repository = "https://github.com/totonga/odsbox-jaquel-mcp"
Tracker = "https://github.com/totonga/odsbox-jaquel-mcp/issues"

[tool.black]
line-length = 119
target-version = ["py313"]

[tool.setuptools_scm]
tag_regex = "^v(?P<version>[0-9]+\\.[0-9]+\\.[0-9]+)$"

[tool.isort]
profile = "black"
line_length = 119

[tool.flake8]
max-line-length = 119
extend-ignore = ["E203", "W503", "W293"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
markers = [
    "integration: marks tests as integration tests (requires live ODS server)",
    "unit: marks tests as unit tests",
]
asyncio_mode = "auto"
